refactor: extract calendar components from index.tsx
